refactor(overview): /simplify pass — extract ordinal util + RegionRankChip (#550)
Reuse / quality cleanup surfaced by the /simplify review on the
KpiBulletCard PR:

- Extracted `ordinalSuffix` (suffix-only: "st" / "nd" / …) and
  `toOrdinal` (full ordinal: "1st" / "2nd" / …) to
  `frontend/src/utils/ordinal.ts`. Replaces 4 local copies in:
  - KpiBulletCard.tsx
  - RankingCard.tsx (uses toOrdinal — full form)
  - ClubAnniversaryBadge.tsx (uses ordinalSuffix — suffix only)
  - UpcomingAnniversariesPanel.tsx (uses ordinalSuffix — suffix only)

  The two shapes existed in parallel across the codebase. Naming both
  forms explicitly prevents future "is this just 'st' or '1st'?"
  guesswork.

- Extracted `RegionRankChip` from KpiBulletCard.tsx into its own
  component file. Self-contained (Tooltip + Link + digit extraction)
  and now reusable from any surface that needs region-rank display
  (rankings table, region page, search results).

- Tightened `TierTick.key` to `keyof RecognitionTargets` (was a
  parallel string literal union — now derived from the source type).

- "Targets unavailable" fallback gains `role="status"` and a
  describing `aria-label` so screen readers don't read silence where
  the progressbar would be.

- Inlined the duplicated `<span aria-hidden>·</span>` separators as a
  `<DotSep />` helper.

- Dropped narrative comments in KpiBulletCard and the test file's
  migration-history header docblock — identifiers + structure speak
  for themselves.

Tests: 2295/2295 green.
