+++ frontend/src/components/ClubAnniversaryBadge.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import type { ClubAnniversary } from '../utils/clubAnniversary'
+import { ordinalSuffix } from '../utils/ordinal'
 
 /* ClubAnniversaryBadge (#445) — anniversary display for the club hero.
    Three visual modes:
@@ -15,16 +16,6 @@ export interface ClubAnniversaryBadgeProps {
   anniversary: ClubAnniversary | null
   /** Charter date in display form for the tooltip. */
   charterDateLabel?: string
-}
-
-const ordinalSuffix = (n: number): string => {
-  const mod100 = n % 100
-  if (mod100 >= 11 && mod100 <= 13) return 'th'
-  const mod10 = n % 10
-  if (mod10 === 1) return 'st'
-  if (mod10 === 2) return 'nd'
-  if (mod10 === 3) return 'rd'
-  return 'th'
 }
 
 export const ClubAnniversaryBadge: React.FC<ClubAnniversaryBadgeProps> = ({
