+++ frontend/src/components/KpiBulletCard.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
-import { Link } from 'react-router-dom'
 import { Tooltip, InfoIcon } from './Tooltip'
+import { RegionRankChip } from './RegionRankChip'
+import { toOrdinal } from '../utils/ordinal'
 import type { MetricRankings, RecognitionTargets } from '../types/districts'
 
 export interface KpiBulletCardProps {
@@ -10,72 +11,36 @@ export interface KpiBulletCardProps {
   targets: RecognitionTargets | null
   tooltipContent?: string
   /**
-   * Accessible name for the bullet bar. When omitted, falls back to a
-   * generated label that includes the card title and the Smedley scale.
+   * Accessible name for the bullet bar. Falls back to a generated label
+   * that includes the card title and the Smedley scale.
    */
   barAriaLabel?: string
 }
 
+type TierKey = keyof RecognitionTargets
+
 interface TierTick {
-  key: 'distinguished' | 'select' | 'presidents' | 'smedley'
+  key: TierKey
   shortLabel: string
   value: number
 }
 
-function ordinalSuffix(n: number): string {
-  const mod100 = n % 100
-  if (mod100 >= 11 && mod100 <= 13) return `${n}th`
-  const mod10 = n % 10
-  if (mod10 === 1) return `${n}st`
-  if (mod10 === 2) return `${n}nd`
-  if (mod10 === 3) return `${n}rd`
-  return `${n}th`
-}
+const TIERS: { key: TierKey; shortLabel: string }[] = [
+  { key: 'distinguished', shortLabel: 'D' },
+  { key: 'select', shortLabel: 'S' },
+  { key: 'presidents', shortLabel: 'P' },
+  { key: 'smedley', shortLabel: 'Sm' },
+]
 
 function formatPercentile(percentile: number | null): string {
   if (percentile === null) return '—'
   const rankPercent = Math.round(100 - percentile)
-  return `${ordinalSuffix(rankPercent)} percentile`
+  return `${toOrdinal(rankPercent)} percentile`
 }
 
 function formatPct(n: number): string {
   if (n >= 100) return '100%'
   return `${n.toFixed(2)}%`
-}
-
-const RegionRankChip: React.FC<{
-  region: string
-  regionRank: number | null
-}> = ({ region, regionRank }) => {
-  const digits = String(region).match(/\d+/)?.[0]
-  const tooltipContent =
-    regionRank !== null
-      ? `District's rank within ${region} region (1 = best)`
-      : `Regional ranking data unavailable for ${region}`
-  const label = (
-    <>
-      {region}: {regionRank !== null ? `#${regionRank}` : '—'}
-    </>
-  )
-  const sharedClasses = 'px-2 py-1 rounded-sm bg-gray-100 text-gray-700'
-  return (
-    <Tooltip content={tooltipContent}>
-      {digits ? (
-        <Link
-          to={`/region/${digits}`}
-          className={`${sharedClasses} hover:bg-gray-200`}
-          data-testid="region-rank"
-          aria-label={`View Region ${digits} overview`}
-        >
-          {label}
-        </Link>
-      ) : (
-        <span className={sharedClasses} data-testid="region-rank">
-          {label}
-        </span>
-      )}
-    </Tooltip>
-  )
 }
 
 interface BulletBarProps {
@@ -85,13 +50,6 @@ interface BulletBarProps {
   ariaLabel: string | undefined
 }
 
-const TIERS: Pick<TierTick, 'key' | 'shortLabel'>[] = [
-  { key: 'distinguished', shortLabel: 'D' },
-  { key: 'select', shortLabel: 'S' },
-  { key: 'presidents', shortLabel: 'P' },
-  { key: 'smedley', shortLabel: 'Sm' },
-]
-
 const BulletBar: React.FC<BulletBarProps> = ({
   current,
   targets,
@@ -99,8 +57,7 @@ const BulletBar: React.FC<BulletBarProps> = ({
   ariaLabel,
 }) => {
   const max = targets.smedley
-  const markerRatio = max > 0 ? current / max : 0
-  const markerPct = Math.min(100, markerRatio * 100)
+  const markerPct = Math.min(100, max > 0 ? (current / max) * 100 : 0)
   const allAchieved = current >= max
 
   const tiers: TierTick[] = TIERS.map(t => ({ ...t, value: targets[t.key] }))
@@ -117,7 +74,6 @@ const BulletBar: React.FC<BulletBarProps> = ({
         `${title} — current ${current.toLocaleString()} of ${max.toLocaleString()} (Smedley tier)`
       }
     >
-      {/* Current-value marker — value label + downward triangle above the bar */}
       <div
         data-testid="current-marker"
         data-all-achieved={allAchieved ? 'true' : 'false'}
@@ -135,7 +91,6 @@ const BulletBar: React.FC<BulletBarProps> = ({
         </span>
       </div>
 
-      {/* The bar itself */}
       <div className="relative h-2 rounded-full bg-gray-200">
         <div
           className="absolute inset-y-0 left-0 rounded-full bg-tm-loyal-blue"
@@ -143,7 +98,6 @@ const BulletBar: React.FC<BulletBarProps> = ({
         />
       </div>
 
-      {/* Tier ticks — short vertical line + short label + threshold value */}
       {tiers.map(t => {
         const pos = (t.value / max) * 100
         const achieved = current >= t.value
@@ -169,10 +123,16 @@ const BulletBar: React.FC<BulletBarProps> = ({
   )
 }
 
+const DotSep: React.FC = () => (
+  <span aria-hidden="true" className="text-gray-400">
+    ·
+  </span>
+)
+
 /**
- * KpiBulletCard — single-metric KPI card with a bullet-bar tier-progress
- * visualization. One bar carries the current-value marker plus four tier
- * ticks (D / S / P / Sm) showing the Distinguished District thresholds.
+ * Single-metric KPI card with a bullet-bar tier-progress visualization.
+ * One bar carries the current-value marker plus four tier ticks
+ * (D / S / P / Sm) showing Distinguished District thresholds.
  */
 export const KpiBulletCard: React.FC<KpiBulletCardProps> = ({
   title,
@@ -182,13 +142,24 @@ export const KpiBulletCard: React.FC<KpiBulletCardProps> = ({
   tooltipContent,
   barAriaLabel,
 }) => {
+  const worldRankLabel =
+    rankings.worldRank !== null && rankings.totalDistricts > 0
+      ? `#${rankings.worldRank} of ${rankings.totalDistricts}`
+      : '—'
+  const worldRankTooltip =
+    rankings.worldRank !== null
+      ? "District's rank among all districts worldwide (1 = best)"
+      : 'Ranking data unavailable - district may not have sufficient data for ranking'
+  const percentileTooltip =
+    rankings.worldPercentile !== null
+      ? 'Percentage of districts this district outperforms worldwide'
+      : 'Percentile data unavailable - requires world rank data'
+
   return (
     <div
       className="rounded-lg border border-gray-200 bg-white p-4"
       data-testid="kpi-bullet-card"
     >
-      {/* Header: title (+ optional info tooltip) on the left,
-          big number underneath. */}
       <div className="flex items-center gap-1">
         <h3 className="text-sm font-medium text-gray-700">{title}</h3>
         {tooltipContent && (
@@ -210,40 +181,19 @@ export const KpiBulletCard: React.FC<KpiBulletCardProps> = ({
         {current.toLocaleString()}
       </p>
 
-      {/* Inline rank line: #rank of N · percentile · region rank */}
       <div className="mt-2 flex flex-wrap items-center gap-1.5 text-xs text-gray-600">
-        <Tooltip
-          content={
-            rankings.worldRank !== null
-              ? "District's rank among all districts worldwide (1 = best)"
-              : 'Ranking data unavailable - district may not have sufficient data for ranking'
-          }
-        >
-          <span data-testid="world-rank">
-            {rankings.worldRank !== null && rankings.totalDistricts > 0
-              ? `#${rankings.worldRank} of ${rankings.totalDistricts}`
-              : '—'}
-          </span>
+        <Tooltip content={worldRankTooltip}>
+          <span data-testid="world-rank">{worldRankLabel}</span>
         </Tooltip>
-        <span aria-hidden="true" className="text-gray-400">
-          ·
-        </span>
-        <Tooltip
-          content={
-            rankings.worldPercentile !== null
-              ? 'Percentage of districts this district outperforms worldwide'
-              : 'Percentile data unavailable - requires world rank data'
-          }
-        >
+        <DotSep />
+        <Tooltip content={percentileTooltip}>
           <span data-testid="world-percentile">
             {formatPercentile(rankings.worldPercentile)}
           </span>
         </Tooltip>
         {rankings.region && (
           <>
-            <span aria-hidden="true" className="text-gray-400">
-              ·
-            </span>
+            <DotSep />
             <RegionRankChip
               region={rankings.region}
               regionRank={rankings.regionRank}
@@ -252,7 +202,6 @@ export const KpiBulletCard: React.FC<KpiBulletCardProps> = ({
         )}
       </div>
 
-      {/* Bullet bar OR fallback when targets unavailable */}
       {targets ? (
         <BulletBar
           current={current}
@@ -263,6 +212,8 @@ export const KpiBulletCard: React.FC<KpiBulletCardProps> = ({
       ) : (
         <div
           data-testid="targets-unavailable"
+          role="status"
+          aria-label={`${title} — recognition targets unavailable`}
           className="mt-3 text-xs text-gray-500"
         >
           Targets unavailable
