+++ frontend/src/components/RankingCard.tsx
@@ -1,16 +1,6 @@
 import React from 'react'
 import { Card } from './ui/Card'
-
-/** Format number with ordinal suffix (1st, 2nd, 3rd, 4th, ...) */
-function ordinalSuffix(n: number): string {
-  const mod100 = n % 100
-  if (mod100 >= 11 && mod100 <= 13) return `${n}th`
-  const mod10 = n % 10
-  if (mod10 === 1) return `${n}st`
-  if (mod10 === 2) return `${n}nd`
-  if (mod10 === 3) return `${n}rd`
-  return `${n}th`
-}
+import { toOrdinal } from '../utils/ordinal'
 
 /**
  * Color scheme configuration for RankingCard
@@ -98,7 +88,7 @@ const RankingCard: React.FC<RankingCardProps> = ({
 
   // Format percentile as ordinal (e.g., "12th percentile") (#305)
   const rounded = Math.round(percentile)
-  const percentileDisplay = `${ordinalSuffix(rounded)} percentile`
+  const percentileDisplay = `${toOrdinal(rounded)} percentile`
 
   if (isLoading) {
     return (
