+++ frontend/src/components/UpcomingAnniversariesPanel.tsx
@@ -6,6 +6,7 @@ import {
   isMilestoneYear,
   type ClubAnniversary,
 } from '../utils/clubAnniversary'
+import { ordinalSuffix } from '../utils/ordinal'
 
 /* UpcomingAnniversariesPanel (#446 / #511) — district-level recognition
    planner. Tight, dense rows of clubs whose next anniversary is within
@@ -24,16 +25,6 @@ export interface UpcomingAnniversariesPanelProps {
 interface Row {
   club: ClubTrend
   anniversary: ClubAnniversary
-}
-
-const ordinalSuffix = (n: number): string => {
-  const mod100 = n % 100
-  if (mod100 >= 11 && mod100 <= 13) return 'th'
-  const mod10 = n % 10
-  if (mod10 === 1) return 'st'
-  if (mod10 === 2) return 'nd'
-  if (mod10 === 3) return 'rd'
-  return 'th'
 }
 
 const buildRows = (
